mcp/pyproject.toml
[project]
name = "claude-todo-mcp"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.3",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "redis[asyncio]>=5.2",
    "httpx>=0.28",
    "pydantic-settings>=2.7",
    "orjson>=3.10",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "respx>=0.21",
    "fakeredis>=2.26",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
